Make AI check what it is building while checking for BuildLimits
For building queues. I haven't seen them passing the limit much for unit
queues anyway, not sure if this is necessary there.

They still sometimes pass the limits with 1 or 2. Not seen over that for
now. Without this i remember seeing AI building 5+ palace/strategy
center/propaganda center, so it helps.